// routes.gen.ts
//AUTOMATICALLY GENERATED CODE, regenerates everytime any RoutesGenerator is run 
import * as $0 from "../routes/index.tsx";
import * as $1 from "../routes/auth/login.tsx";
import * as $2 from "../routes/auth/:user.tsx";

const manifest = { 
	routes: [
		{ urlPath: "/", module: $0, relativePath: "../routes/index.tsx"}, 
		{ urlPath: "/auth/login", module: $1, relativePath: "../routes/auth/login.tsx"}, 
		{ urlPath: "/auth/:user", module: $2, relativePath: "../routes/auth/:user.tsx"}, 
	], 
	baseUrl: import.meta.url,
} 

export default manifest;